requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
aiofiles==23.2.1
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
